feat: Add a useStore hook for tear-less non-snapshot proxies.
The idea is to avoid the awkwardness around "doing I pass my
children a snapshot or a store" by keeping a "unified proxy"
model (similar to mobx) that can always both get + set.

We still do usage-based tracking, still via proxy-compare, still
with useSyncExternalStore, but instead of comparing prev/curr
snapshots to determine "dirty reads", we compare accessed vs.
mutated paths, and re-render overlap between the two.

This will technically re-render on mutations that flip/flop values,
i.e. foo.a = 1, foo.a = 2, back to foo.a = 1, we won't realize
that `a` has not _actually_ changed since our last render, but
this seems like a minor issue for the upshots of unified proxy
simplicity, still usage-based tracking, and avoiding the cost
of snapshots.

(The insight is that, afaiu since useSyncExternalStore forces a
re-render before paint on an inconsistency anyway, we don't
really need an immutable snapshot, i.e. it's not as if React
would let the now-stable-but-technically-consistent render
complete, in which case it'd make since to keep the snapshot
to let the render finish. But if our interrupted render's
partial results get thrown away anyway, we might as well
always render directly from the mutable proxy.)